requirements.txt
Pillow==2.4.0
distribute==0.6.49
ipython==2.0.0
stevedore==0.10
virtualenv==1.9.1
virtualenv-clone==0.2.4
virtualenvwrapper==4.0
wsgiref==0.1.2
